Removed chocolatey installation instructions from the official docs
There have been ~20k installations via chocolatey of all versions of Material,
which is what is installed through `pip` in 3 days. The usage is very, very low
and also not recommended, because it will very probably have problems with extensions,
plugins etc. For this reason, this commit removes the instructions from the docs.